+++ LondonDataServices.IDecide.Core.Tests.Unit/Services/Foundations/DecisionTypes/DecisionTypeServiceTests.Exceptions.RemoveById.cs
@@ -1,0 +1,62 @@
+﻿// ---------------------------------------------------------
+// Copyright (c) North East London ICB. All rights reserved.
+// ---------------------------------------------------------
+
+using System;
+using System.Threading.Tasks;
+using FluentAssertions;
+using LondonDataServices.IDecide.Core.Models.Foundations.DecisionTypes;
+using LondonDataServices.IDecide.Core.Models.Foundations.DecisionTypes.Exceptions;
+using Moq;
+
+namespace LondonDataServices.IDecide.Core.Tests.Unit.Services.Foundations.DecisionTypes
+{
+    public partial class DecisionTypeServiceTests
+    {
+        [Fact]
+        public async Task ShouldThrowValidationExceptionOnRemoveIfIdIsInvalidAndLogItAsync()
+        {
+            // given
+            Guid invalidDecisionTypeId = Guid.Empty;
+
+            var invalidDecisionTypeException =
+                new InvalidDecisionTypeException(
+                    message: "Invalid decision type. Please correct the errors and try again.");
+
+            invalidDecisionTypeException.AddData(
+                key: nameof(DecisionType.Id),
+                values: "Id is invalid");
+
+            var expectedDecisionTypeValidationException =
+                new DecisionTypeValidationException(
+                    message: "DecisionType validation error occurred, please fix errors and try again.",
+                    innerException: invalidDecisionTypeException);
+
+            // when
+            ValueTask<DecisionType> removeDecisionTypeByIdTask =
+                this.decisionTypeService.RemoveDecisionTypeByIdAsync(invalidDecisionTypeId);
+
+            DecisionTypeValidationException actualDecisionTypeValidationException =
+                await Assert.ThrowsAsync<DecisionTypeValidationException>(
+                    removeDecisionTypeByIdTask.AsTask);
+
+            // then
+            actualDecisionTypeValidationException.Should()
+                .BeEquivalentTo(expectedDecisionTypeValidationException);
+
+            this.loggingBrokerMock.Verify(broker =>
+                broker.LogErrorAsync(It.Is(SameExceptionAs(
+                    expectedDecisionTypeValidationException))),
+                        Times.Once);
+
+            this.storageBrokerMock.Verify(broker =>
+                broker.DeleteDecisionTypeAsync(It.IsAny<DecisionType>()),
+                    Times.Never);
+
+            this.dateTimeBrokerMock.VerifyNoOtherCalls();
+            this.securityBrokerMock.VerifyNoOtherCalls();
+            this.loggingBrokerMock.VerifyNoOtherCalls();
+            this.storageBrokerMock.VerifyNoOtherCalls();
+        }
+    }
+}
