+++ LondonDataServices.IDecide.Core.Tests.Unit/Services/Foundations/DecisionTypes/DecisionTypeServiceTests.Logic.RemoveById.cs
@@ -1,0 +1,57 @@
+﻿// ---------------------------------------------------------
+// Copyright (c) North East London ICB. All rights reserved.
+// ---------------------------------------------------------
+
+using System;
+using System.Threading.Tasks;
+using FluentAssertions;
+using Force.DeepCloner;
+using LondonDataServices.IDecide.Core.Models.Foundations.DecisionTypes;
+using Moq;
+
+namespace LondonDataServices.IDecide.Core.Tests.Unit.Services.Foundations.DecisionTypes
+{
+    public partial class DecisionTypeServiceTests
+    {
+        [Fact]
+        public async Task ShouldRemoveDecisionTypeByIdAsync()
+        {
+            // given
+            Guid randomId = Guid.NewGuid();
+            Guid inputDecisionTypeId = randomId;
+            DecisionType randomDecisionType = CreateRandomDecisionType();
+            DecisionType storageDecisionType = randomDecisionType;
+            DecisionType expectedInputDecisionType = storageDecisionType;
+            DecisionType deletedDecisionType = expectedInputDecisionType;
+            DecisionType expectedDecisionType = deletedDecisionType.DeepClone();
+
+            this.storageBrokerMock.Setup(broker =>
+                broker.SelectDecisionTypeByIdAsync(inputDecisionTypeId))
+                    .ReturnsAsync(storageDecisionType);
+
+            this.storageBrokerMock.Setup(broker =>
+                broker.DeleteDecisionTypeAsync(expectedInputDecisionType))
+                    .ReturnsAsync(deletedDecisionType);
+
+            // when
+            DecisionType actualDecisionType =
+                await this.decisionTypeService.RemoveDecisionTypeByIdAsync(inputDecisionTypeId);
+
+            // then
+            actualDecisionType.Should().BeEquivalentTo(expectedDecisionType);
+
+            this.storageBrokerMock.Verify(broker =>
+                broker.SelectDecisionTypeByIdAsync(inputDecisionTypeId),
+                    Times.Once);
+
+            this.storageBrokerMock.Verify(broker =>
+                broker.DeleteDecisionTypeAsync(expectedInputDecisionType),
+                    Times.Once);
+
+            this.dateTimeBrokerMock.VerifyNoOtherCalls();
+            this.securityBrokerMock.VerifyNoOtherCalls();
+            this.storageBrokerMock.VerifyNoOtherCalls();
+            this.loggingBrokerMock.VerifyNoOtherCalls();
+        }
+    }
+}
