+++ LondonDataServices.IDecide.Core.Tests.Unit/Services/Foundations/DecisionTypes/DecisionTypeServiceTests.Validations.RemoveById.cs
@@ -1,0 +1,224 @@
+﻿// ---------------------------------------------------------
+// Copyright (c) North East London ICB. All rights reserved.
+// ---------------------------------------------------------
+
+using System;
+using System.Threading.Tasks;
+using FluentAssertions;
+using LondonDataServices.IDecide.Core.Models.Foundations.DecisionTypes;
+using LondonDataServices.IDecide.Core.Models.Foundations.DecisionTypes.Exceptions;
+using Microsoft.Data.SqlClient;
+using Microsoft.EntityFrameworkCore;
+using Moq;
+
+namespace LondonDataServices.IDecide.Core.Tests.Unit.Services.Foundations.DecisionTypes
+{
+    public partial class DecisionTypeServiceTests
+    {
+        [Fact]
+        public async Task ShouldThrowCriticalDependencyExceptionOnRemoveIfSqlErrorOccursAndLogItAsync()
+        {
+            // given
+            DecisionType randomDecisionType = CreateRandomDecisionType();
+            SqlException sqlException = GetSqlException();
+
+            var failedDecisionTypeStorageException =
+                new FailedDecisionTypeStorageException(
+                    message: "Failed decision type storage error occurred, contact support.",
+                    innerException: sqlException);
+
+            var expectedDecisionTypeDependencyException =
+                new DecisionTypeDependencyException(
+                    message: "DecisionType dependency error occurred, contact support.",
+                    innerException: failedDecisionTypeStorageException);
+
+            this.storageBrokerMock.Setup(broker =>
+                broker.SelectDecisionTypeByIdAsync(randomDecisionType.Id))
+                    .ThrowsAsync(sqlException);
+
+            // when
+            ValueTask<DecisionType> addDecisionTypeTask =
+                this.decisionTypeService.RemoveDecisionTypeByIdAsync(randomDecisionType.Id);
+
+            DecisionTypeDependencyException actualDecisionTypeDependencyException =
+                await Assert.ThrowsAsync<DecisionTypeDependencyException>(
+                    addDecisionTypeTask.AsTask);
+
+            // then
+            actualDecisionTypeDependencyException.Should()
+                .BeEquivalentTo(expectedDecisionTypeDependencyException);
+
+            this.storageBrokerMock.Verify(broker =>
+                broker.SelectDecisionTypeByIdAsync(randomDecisionType.Id),
+                    Times.Once);
+
+            this.loggingBrokerMock.Verify(broker =>
+                broker.LogCriticalAsync(It.Is(SameExceptionAs(
+                    expectedDecisionTypeDependencyException))),
+                        Times.Once);
+
+            this.storageBrokerMock.Verify(broker =>
+                broker.DeleteDecisionTypeAsync(It.IsAny<DecisionType>()),
+                    Times.Never);
+
+            this.dateTimeBrokerMock.Verify(broker =>
+                broker.GetCurrentDateTimeOffsetAsync(),
+                    Times.Never);
+
+            this.dateTimeBrokerMock.VerifyNoOtherCalls();
+            this.securityBrokerMock.VerifyNoOtherCalls();
+            this.storageBrokerMock.VerifyNoOtherCalls();
+            this.loggingBrokerMock.VerifyNoOtherCalls();
+        }
+
+        [Fact]
+        public async Task ShouldThrowDependencyValidationOnRemoveIfDatabaseUpdateConcurrencyErrorOccursAndLogItAsync()
+        {
+            // given
+            Guid someDecisionTypeId = Guid.NewGuid();
+
+            var databaseUpdateConcurrencyException =
+                new DbUpdateConcurrencyException();
+
+            var lockedDecisionTypeException =
+                new LockedDecisionTypeException(
+                    message: "Locked decision type record exception, please fix errors and try again.",
+                    innerException: databaseUpdateConcurrencyException);
+
+            var expectedDecisionTypeDependencyValidationException =
+                new DecisionTypeDependencyValidationException(
+                    message: "DecisionType dependency validation occurred, please fix errors and try again.",
+                    innerException: lockedDecisionTypeException);
+
+            this.storageBrokerMock.Setup(broker =>
+                broker.SelectDecisionTypeByIdAsync(It.IsAny<Guid>()))
+                    .ThrowsAsync(databaseUpdateConcurrencyException);
+
+            // when
+            ValueTask<DecisionType> removeDecisionTypeByIdTask =
+                this.decisionTypeService.RemoveDecisionTypeByIdAsync(someDecisionTypeId);
+
+            DecisionTypeDependencyValidationException actualDecisionTypeDependencyValidationException =
+                await Assert.ThrowsAsync<DecisionTypeDependencyValidationException>(
+                    removeDecisionTypeByIdTask.AsTask);
+
+            // then
+            actualDecisionTypeDependencyValidationException.Should()
+                .BeEquivalentTo(expectedDecisionTypeDependencyValidationException);
+
+            this.storageBrokerMock.Verify(broker =>
+                broker.SelectDecisionTypeByIdAsync(It.IsAny<Guid>()),
+                    Times.Once);
+
+            this.loggingBrokerMock.Verify(broker =>
+                broker.LogErrorAsync(It.Is(SameExceptionAs(
+                    expectedDecisionTypeDependencyValidationException))),
+                        Times.Once);
+
+            this.storageBrokerMock.Verify(broker =>
+                broker.DeleteDecisionTypeAsync(It.IsAny<DecisionType>()),
+                    Times.Never);
+
+            this.dateTimeBrokerMock.VerifyNoOtherCalls();
+            this.securityBrokerMock.VerifyNoOtherCalls();
+            this.storageBrokerMock.VerifyNoOtherCalls();
+            this.loggingBrokerMock.VerifyNoOtherCalls();
+        }
+
+        [Fact]
+        public async Task ShouldThrowDependencyExceptionOnRemoveWhenSqlExceptionOccursAndLogItAsync()
+        {
+            // given
+            Guid someDecisionTypeId = Guid.NewGuid();
+            SqlException sqlException = GetSqlException();
+
+            var failedDecisionTypeStorageException =
+                new FailedDecisionTypeStorageException(
+                    message: "Failed decision type storage error occurred, contact support.",
+                    innerException: sqlException);
+
+            var expectedDecisionTypeDependencyException =
+                new DecisionTypeDependencyException(
+                    message: "DecisionType dependency error occurred, contact support.",
+                    innerException: failedDecisionTypeStorageException);
+
+            this.storageBrokerMock.Setup(broker =>
+                broker.SelectDecisionTypeByIdAsync(It.IsAny<Guid>()))
+                    .ThrowsAsync(sqlException);
+
+            // when
+            ValueTask<DecisionType> deleteDecisionTypeTask =
+                this.decisionTypeService.RemoveDecisionTypeByIdAsync(someDecisionTypeId);
+
+            DecisionTypeDependencyException actualDecisionTypeDependencyException =
+                await Assert.ThrowsAsync<DecisionTypeDependencyException>(
+                    deleteDecisionTypeTask.AsTask);
+
+            // then
+            actualDecisionTypeDependencyException.Should()
+                .BeEquivalentTo(expectedDecisionTypeDependencyException);
+
+            this.storageBrokerMock.Verify(broker =>
+                broker.SelectDecisionTypeByIdAsync(It.IsAny<Guid>()),
+                    Times.Once);
+
+            this.loggingBrokerMock.Verify(broker =>
+                broker.LogCriticalAsync(It.Is(SameExceptionAs(
+                    expectedDecisionTypeDependencyException))),
+                        Times.Once);
+
+            this.dateTimeBrokerMock.VerifyNoOtherCalls();
+            this.securityBrokerMock.VerifyNoOtherCalls();
+            this.storageBrokerMock.VerifyNoOtherCalls();
+            this.loggingBrokerMock.VerifyNoOtherCalls();
+        }
+
+        [Fact]
+        public async Task ShouldThrowServiceExceptionOnRemoveIfServiceErrorOccursAndLogItAsync()
+        {
+            // given
+            Guid someDecisionTypeId = Guid.NewGuid();
+            var serviceException = new Exception();
+
+            var failedDecisionTypeServiceException =
+                new FailedDecisionTypeServiceException(
+                    message: "Failed decision type service occurred, please contact support",
+                    innerException: serviceException);
+
+            var expectedDecisionTypeServiceException =
+                new DecisionTypeServiceException(
+                    message: "DecisionType service error occurred, contact support.",
+                    innerException: failedDecisionTypeServiceException);
+
+            this.storageBrokerMock.Setup(broker =>
+                broker.SelectDecisionTypeByIdAsync(It.IsAny<Guid>()))
+                    .ThrowsAsync(serviceException);
+
+            // when
+            ValueTask<DecisionType> removeDecisionTypeByIdTask =
+                this.decisionTypeService.RemoveDecisionTypeByIdAsync(someDecisionTypeId);
+
+            DecisionTypeServiceException actualDecisionTypeServiceException =
+                await Assert.ThrowsAsync<DecisionTypeServiceException>(
+                    removeDecisionTypeByIdTask.AsTask);
+
+            // then
+            actualDecisionTypeServiceException.Should()
+                .BeEquivalentTo(expectedDecisionTypeServiceException);
+
+            this.storageBrokerMock.Verify(broker =>
+                broker.SelectDecisionTypeByIdAsync(It.IsAny<Guid>()),
+                        Times.Once());
+
+            this.loggingBrokerMock.Verify(broker =>
+                broker.LogErrorAsync(It.Is(SameExceptionAs(
+                    expectedDecisionTypeServiceException))),
+                        Times.Once);
+
+            this.dateTimeBrokerMock.VerifyNoOtherCalls();
+            this.securityBrokerMock.VerifyNoOtherCalls();
+            this.storageBrokerMock.VerifyNoOtherCalls();
+            this.loggingBrokerMock.VerifyNoOtherCalls();
+        }
+    }
+}
